requirements.txt
# Core Django Framework
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.4

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests and OAuth
requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2

# Image Processing
Pillow==10.4.0

# Development and Testing (optional)
# Uncomment these if you want to include development dependencies
# pytest==8.3.4
# pytest-django==4.9.0
# coverage==7.6.7

